fix ubsan warning
interp.c:2490:12: runtime error: left shift of 1048575 by 12 places cannot be represented in type 'int'

+++ src/include/emu-ldt.h
@@ -27,7 +27,7 @@
 
 #define MAX_SELECTORS	LDT_ENTRIES
 
-#define DT_LIMIT(dp)		(((dp)->limit_hi<<16) | ((dp)->limit_lo))
+#define DT_LIMIT(dp)		((uint32_t)((dp)->limit_hi<<16) | ((dp)->limit_lo))
 #define DT_BASE(dp)		(((dp)->base_hi<<24) | ((dp)->base_mid<<16) | ((dp)->base_lo))
 #if defined(i386)||defined(__i386)||defined(__i386__)||defined(__x86_64__)
 #define DT_FLAGS(dp)		(*((unsigned short *)(((char *)(dp))+5))&0xf0ff)
